+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Bugs
 
 #### Improvements
+* Fix #7522: improve dependency management for kubernetes-httpclient-okhttp
 
 #### Dependency Upgrade